created quiz_results.ejs file,  quiz_results.scss file, and added query to routes/quiz_results.js file
