--- routes/attempts.js
@@ -1,28 +1,0 @@
-/*
- * All routes for Widgets are defined here
- * Since this file is loaded in server.js into api/widgets,
- *   these routes are mounted onto /widgets
- * See: https://expressjs.com/en/guide/using-middleware.html#middleware.router
- */
-
-const express = require('express');
-const router  = express.Router();
-
-module.exports = (db) => {
-  router.get("/", (req, res) => {
-    let query = `SELECT * FROM attempts`;
-    console.log(query);
-    db.query(query)
-      .then(data => {
-        const attempts = data.rows;
-        console.log(attempts);
-        res.json({ attempts });
-      })
-      .catch(err => {
-        res
-          .status(500)
-          .json({ error: err.message });
-      });
-  });
-  return router;
-};
