+++ server.js
@@ -38,12 +38,14 @@ app.use(express.static("public"));
 const usersRoutes = require("./routes/users");
 const widgetsRoutes = require("./routes/widgets");
 const quizzesRoutes = require("./routes/quizzes");
+const quizResultsRoutes = require("./routes/quiz_results")
 
 // Mount all resource routes
 // Note: Feel free to replace the example routes below with your own
 app.use("/api/users", usersRoutes(db));
 app.use("/api/widgets", widgetsRoutes(db));
 app.use("/api/quizzes", quizzesRoutes(db));
+app.use("/api/quizzes/results", quizResultsRoutes(db));
 // Note: mount other resources here, using the same pattern above
 
 // Home page
