+++ styles/main.scss
@@ -34,7 +34,6 @@ td {
   padding: 5px;
 }
 
-
 .create_new_quiz{
   background-color: C7E4FA;
 }
