+++ styles/quiz_results.scss
@@ -1,0 +1,11 @@
+.quiz_results {
+  background: #FAC7D9;
+  margin-top: 10%;
+  margin-bottom: 5%;
+  padding-top: 1em;
+  height: 612px;
+  width: 80%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
