+++ views/quiz_results.ejs
@@ -1,0 +1,75 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <title>Home Page</title>
+
+  <link rel="stylesheet" href="/vendor/normalize-4.1.1.css" type="text/css" />
+  <link rel="stylesheet" href="/vendor/border-box.css" type="text/css" />
+  <link rel="stylesheet" href="/styles/main.css" type="text/css" />
+  <link rel="stylesheet" href="/styles/layout.css" type="text/css" />
+  <link rel="stylesheet" href="/styles/nav.css" type="text/css" />
+
+  <script type="text/javascript" src="/vendor/jquery-3.0.0.js"></script>
+  <script type="text/javascript" src="/scripts/app.js"></script>
+
+  <!-- Font awesome library-->
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css" type="text/css" />
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.4/css/all.min.css" />
+
+  <!-- fonts for CSS -->
+  <link rel="preconnect" href="https://fonts.googleapis.com">
+<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+<link href="https://fonts.googleapis.com/css2?family=Roboto:wght@100&display=swap" rel="stylesheet">
+
+</head>
+
+
+
+<body>
+  <nav>
+    <span class = "bold">LOGO</span>
+    <div class = "links_header">
+      <a>Quizzes</a>
+      <a>Subjects</a>
+      <a>Create</a>
+      <a>Explore</a>
+    </div>
+    <!-- <a>Logged in as User</a>
+    <a>Log Out</a> -->
+    <a> login</a>
+  </nav>
+<!--
+  <h1>Have fun! 🥔</h1> -->
+  <!-- <div class = "create_new_quiz"> -->
+    <!-- <p>Create a new quiz</p> -->
+    <!-- <form>
+      <label>Quiz</label>
+      <input name = "Quiz name" required/>
+    </form> -->
+    
+  <!-- </div> -->
+
+  <div class = "quiz_results">
+    <table>
+      <thead>
+        <tr>
+          <th scope="col:">Question</th>
+          <th scope="col">Answer</th>
+          <th scope="col">Score</th>
+        </tr>
+      </thead>
+      <tbody>        
+        <% for(let item of results) { %>
+          <tr>
+            <td><%= item.question %><td>
+            <td><%= item.answer %></td>
+            <td><%= item.correct %></td>
+          </tr>
+        <% } %>
+      </tbody>
+    </table>
+  </div>
+</body>
+
+</html>
